Better exit codes, jobs finished w error/cancelled
Fixes #8
Error codes have been partially rewritten an reorganized, some has been
converted into warnings.
The priority is parsing the JSON file that contains the pre-processed
data. If both files are retrieved but the job was completed with error
or cancelled, the corresponding exit code is returned. In order to know
this, the scheduler plugin writes a `JobInfo.annotation` (since the
job_state alone is "FINISHED" both for completed and cancelled jobs),
that is then read by the parser.