Avoid parsing of scheduler stdout & stderr that raised errors
There are no stdout and stderr files returned by tomato, so we do not
need to parse them (which would throw an error).
Still '_scheduler-stderr.txt' and '_scheduler-stdout.txt' are in the
list of files to be retrieved. I do not know if there is a way to remove
them without breaking aiida-core.