(PR #19) Add cycling sequence workchain
This PR adds a workflow allowing users to couple multiple protocols per battery sample in an experiment, with separate settings and monitors defined for each protocol. The main motivation of the PR is to ensure protocols run in the requested sequence.

The workflow includes cross validation ensuring that settings/monitors match selected protocols. In addition, a check is applied between protocols ensuring previous protocols finished successfully, terminating the sequence otherwise.