(PR #37) Updating the schemas and cycling workchain
The battery schema has been updated to account for most, if not all robot output fields. This was primarily motivated by the need for electrode weights during visualization (capacity normalization), then expanded to account for all fields in case needed in the future (and also, for good measure). The cycling workchain was adjusted accordingly. 

Unrelated to the schemas but also included here, the new group label field was added to the workchain to support auto-generating experiment groups on submission.